minor updates to infra stack
