+++ src/Templates/content/Company.AppName.Infra/CoreExStack.cs
@@ -3,7 +3,6 @@ using System.Threading.Tasks;
 using Company.AppName.Infra.Services;
 using Pulumi;
 using Pulumi.AzureNative.Resources;
-using AD = Pulumi.AzureAD;
 
 namespace Company.AppName.Infra;
 
@@ -67,6 +66,8 @@ public static class CoreExStack
             Tags = tags
         });
 
+
+
         // Permissions for function app
         storage.AddAccess(apps.FunctionPrincipalId, "functionApp");
         serviceBus.AddAccess(apps.FunctionPrincipalId, "functionApp");
@@ -90,54 +91,5 @@ public static class CoreExStack
             ["FunctionSwaggerUrl"] = apps.FunctionSwaggerUrl,
             ["AppSwaggerUrl"] = apps.AppSwaggerUrl,
         };
-    }
-
-    public class StackConfiguration
-    {
-        public Input<string>? SqlAdAdminLogin { get; private set; }
-        public Input<string>? SqlAdAdminPassword { get; private set; }
-        public bool IsAppsDeploymentEnabled { get; private set; }
-        public bool IsDBSchemaDeploymentEnabled { get; private set; }
-        public string PendingVerificationsQueue { get; private set; } = default!;
-        public string VerificationResultsQueue { get; private set; } = default!;
-        public string MassPublishQueue { get; private set; } = default!;
-
-        private StackConfiguration() { }
-
-        public static async Task<StackConfiguration> CreateConfiguration()
-        {
-            // read stack config
-            var config = new Config();
-
-            // get some info from Azure AD
-            var domainResult = await AD.GetDomains.InvokeAsync(new AD.GetDomainsArgs { OnlyDefault = true });
-            var defaultUsername = $"sqlGlobalAdAdmin{Pulumi.Deployment.Instance.StackName}@{domainResult.Domains[0].DomainName}";
-            var defaultPassword = new Pulumi.Random.RandomPassword("sqlAdPassword", new()
-            {
-                Length = 32,
-                Upper = true,
-                Number = true,
-                Special = true,
-                OverrideSpecial = "@",
-                MinLower = 2,
-                MinUpper = 2,
-                MinSpecial = 2,
-                MinNumeric = 2
-            }).Result;
-
-            Log.Info($"Default username is: {defaultUsername}");
-
-            return new StackConfiguration
-            {
-                SqlAdAdminLogin = Extensions.GetConfigValue("sqlAdAdmin", defaultUsername),
-                SqlAdAdminPassword = Extensions.GetConfigValue("sqlAdPassword", defaultPassword),
-                IsAppsDeploymentEnabled = config.GetBoolean("isAppsDeploymentEnabled") ?? false,
-                IsDBSchemaDeploymentEnabled = config.GetBoolean("isDBSchemaDeploymentEnabled") ?? false,
-
-                PendingVerificationsQueue = config.Get("pendingVerificationsQueue") ?? "pendingVerifications",
-                VerificationResultsQueue = config.Get("verificationResultsQueue") ?? "verificationResults",
-                MassPublishQueue = config.Get("massPublishQueue") ?? "massPublish"
-            };
-        }
     }
 }
